Refactor user profile management and error handling
- Updated import paths for ApiErrorResponse type.
- Removed unused components related to user profile and password management.
- Created new components and hooks for user profile management under account-settings.
- Implemented user payment information page with a placeholder for future functionality.
- Adjusted routing to reflect new account-settings structure.
- Enhanced validation schema for user profile updates.
- Added user navigation hooks for better role-based menu management.
